requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1
_openmp_mutex=5.1
absl-py=1.2.0
alembic=1.8.1
amqp=5.1.1
aniso8601=9.0.1
anyio=3.6.1
apache-airflow=2.3.4
apache-airflow-providers-celery=3.0.0
apache-airflow-providers-common-sql=1.1.0
apache-airflow-providers-ftp=3.1.0
apache-airflow-providers-http=4.0.0
apache-airflow-providers-imap=3.0.0
apache-airflow-providers-sqlite=3.2.0
apispec=3.3.2
argcomplete=2.0.0
arrow=1.2.3
asttokens=2.0.8
astunparse=1.6.3
attrs=22.1.0
babel=2.10.3
beautifulsoup4=4.11.1
billiard=3.6.4.0
bleach=5.0.1
blinker=1.5
boto3=1.24.89
botocore=1.27.89
ca-certificates=2023.05.30
cachelib=0.9.0
cachetools=5.3.0
cattrs=22.1.0
celery=5.2.7
certifi=2022.6.15
charset-normalizer=2.0.12
click=8.1.3
click-didyoumean=0.3.0
click-plugins=1.1.1
click-repl=0.2.0
clickclick=20.10.2
colorama=0.4.5
colorlog=4.8.0
comm=0.1.3
commonmark=0.9.1
configparser=5.3.0
connexion=2.14.0
cron-descriptor=1.2.31
croniter=1.3.5
cryptography=36.0.2
cycler=0.11.0
debugpy=1.6.3
dill=0.3.1.1
dnspython=2.2.1
docutils=0.19
email-validator=1.2.1
entrypoints=0.4
exceptiongroup=1.0.0rc8
executing=0.10.0
fastjsonschema=2.16.1
filelock=3.8.0
flask=2.2.2
flask-appbuilder=4.1.3
flask-babel=2.0.0
flask-caching=2.0.1
flask-jwt-extended=4.4.4
flask-login=0.6.2
flask-session=0.4.0
flask-sqlalchemy=2.5.1
flask-wtf=0.15.1
flatbuffers=22.9.24
flower=1.2.0
fonttools=4.37.1
fqdn=1.5.1
gast=0.4.0
geopandas=0.13.2
google-auth=2.12.0
google-auth-oauthlib=0.5.0
google-pasta=0.2.0
google-cloud-storage
graphene=3.1
graphql-core=3.2.1
graphql-relay=3.2.0
greenlet=1.1.2
grpcio=1.49.1
gunicorn=20.1.0
h11=0.12.0
h5py=3.7.0
httpcore=0.15.0
httpx=0.23.0
huggingface-hub=0.9.1
humanize=4.3.0
idna=3.3
importlib-metadata=4.12.0
importlib-resources=5.9.0
inflection=0.5.1
ipykernel=6.15.1
ipython=8.4.0
ipywidgets=8.0.1
isodate=0.6.1
isoduration=20.11.0
itsdangerous=2.1.2
jax=0.4.8
jedi=0.18.1
jinja2=3.1.2
jmespath=1.0.1
joblib=1.2.0
jsonschema=4.13.0
jupyter=1.0.0
jupyter-client=7.4.4
jupyter-console=6.4.4
jupyter-events=0.6.3
jupyter-server=2.5.0
jupyter-server-terminals=0.4.4
jupyterlab-widgets=3.0.2
keras=2.12.0
keras-preprocessing=1.1.2
kiwisolver=1.4.4
kombu=5.2.4
lazy-object-proxy=1.7.1
ld_impl_linux-64=2.38
libclang=14.0.6
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0
libgomp=11.2.0
libstdcxx-ng=11.2.0
linkify-it-py=2.0.0
lockfile=0.12.2
lxml=4.9.1
mako=1.2.1
markdown=3.4.1
markdown-it-py=2.1.0
markupsafe=2.1.1
marshmallow=3.17.0
marshmallow-enum=1.5.1
marshmallow-oneofschema=3.0.1
marshmallow-sqlalchemy=0.26.1
matplotlib=3.5.3
mdit-py-plugins=0.3.0
mdurl=0.1.2
mistune=2.0.4
ml-dtypes=0.1.0
nbclassic=0.5.6
nbclient=0.6.6
nbconvert=7.0.0
nbformat=5.4.0
ncurses=6.4
nest-asyncio=1.5.5
networkx=2.8.6
nltk=3.8.1
notebook=6.4.12
numpy=1.23.2
oauthlib=3.2.2
openssl=1.1.1t
opt-einsum=3.3.0
packaging=21.3
pandas=1.4.3
pathspec=0.9.0
pendulum=2.1.2
pexpect=4.8.0
pillow=9.2.0
pip=23.0.1
platformdirs=3.5.0
pluggy=1.0.0
prison=0.2.1
prometheus-client=0.14.1
prompt-toolkit=3.0.30
protobuf=3.20.3
psutil=5.9.1
psycopg-binary=3.1.1
pyasn1=0.5.0
pyasn1-modules=0.3.0
pycountry=22.3.5
pygments=2.13.0
pyjwt=2.4.0
pyparsing=3.0.9
pyproj=3.5.0
pyrsistent=0.18.1
python=3.8.10
python-daemon=2.3.1
python-graphviz=0.20.1
python-json-logger=2.0.7
python-nvd3=0.15.0
python-slugify=6.1.2
pytz=2022.1
pytzdata=2020.1
pyyaml=6.0
pyzmq=24.0.0
qtconsole=5.3.1
qtpy=2.2.0
rdflib=6.2.0
readline=8.2
regex=2022.9.13
requests=2.28.0
requests-oauthlib=1.3.1
requests-toolbelt=0.9.1
rfc3339-validator=0.1.4
rfc3986=1.5.0
rfc3986-validator=0.1.1
rich=12.5.1
rsa=4.9
s3transfer=0.6.0
scikit-learn=1.1.2
scipy=1.9.1
sematch=1.0.4
send2trash=1.8.0
sentence-transformers=2.2.2
sentencepiece=0.1.97
setproctitle=1.3.2
setuptools=67.8.0
shapely=2.0.1
six=1.16.0
sniffio=1.2.0
soupsieve=2.3.2.post1
spacy=3.6.1
sparqlwrapper=2.0.0
sqlalchemy=1.4.27
sqlalchemy-jsonfield=1.0.0
sqlalchemy-utils=0.38.3
sqlite=3.41.2
sqlparse=0.4.2
stack-data=0.4.0
swagger-ui-bundle=0.0.9
tabulate=0.8.10
tenacity=8.0.1
tensorboard=2.12.3
tensorboard-data-server=0.7.0
tensorflow=2.12.0
tensorflow-estimator=2.12.0
tensorflow-io-gcs-filesystem=0.32.0
termcolor=1.1.0
terminado=0.15.0
text-unidecode=1.3
threadpoolctl=3.1.0
tinycss2=1.1.1
tk=8.6.12
tokenizers=0.12.1
torch=1.12.1
torchvision=0.13.1
tornado=6.2
tqdm=4.64.1
traitlets=5.6.0
transformers=4.22.2
typing-extensions=4.3.0
uc-micro-py=1.0.1
unicodecsv=0.14.1
uri-template=1.2.0
urllib3=1.26.11
vine=5.0.0
wcwidth=0.2.5
webcolors=1.13
werkzeug=2.2.2
wheel=0.38.4
widgetsnbextension=4.0.2
wrapt=1.14.1
wtforms=2.3.3
xz=5.4.2
zipp=3.8.1
zlib=1.2.13
